Add TailwindCSS
To add TailwindCSS, we first need to add yarn and webpacker.

I added a dependency to our Gemfile on `webpacker`. This is standard in
Rails 6, but we started with a minimal install so we needed to add it.
After adding it, I ran `bin/rails webpacker:install` to set up the
scaffolding.

From there, I was able to follow this guide:
https://davidteren.medium.com/tailwindcss-2-0-with-rails-6-1-postcss-8-0-9645e235892d

For the most part, this involved introducing `application.scss`,
`application.js`, and installing TailwindCSS with Yarn.