Revert "Add TailwindCSS" (But keep Tailwind from CDN)
This reverts commit bac858c23c48ec2b648ef830d4e7c1606a1abb0e.

Tailwind, Webpacker, etc weren't in a good state. To simplify things,
let's just use TailwindCSS from the CDN. This isn't ideal, but removes a
bunch of complexity from our app until we have time to configure it
carefully and correctly.

This commit reverts the commit that added Tailwind, Webpacker, etc, and
instead just uses a link to a single CSS in application.html.erb.