Fix users_controller POST test
To test `users_controller#create` (POST `/users`), we want to test
creating a new user. Since this user will be a new user, we should not
use a user that's already in our test fixtures - that user will already
exist.

Also, we need to pass `password_confirmation` with our params. Code on
the server will validate that `password` matches `password_confirmation`
before creating the new user, and will fail if they don't match or if
one is missing.