Fix users_controller#update test
Similar to the create test - we don't want to pass the information from
the existing fixture here because that doesn't change anything. Instead,
we want to change something to be different than what already exists.

This test was failing because we were passing `password` without
`password_confirmation`. I updated the spec to not pass either for now,
and now the test is passing.