Provide default env config for dev and test
We should make it as easy as possible for a new developer to set up a
dev environment. To that end, we shouldn't require a bunch of env vars
to be set if we don't have a good reason to. For development and test
environments, we can provide a default username and password for
postgres, and we can configure that default user/pass in our
docker-compose file. I've set this up in such a way that any of it can
still be overridden by environment variables if you choose to do so, but
you don't have to -- it will work out of the box.

- Edit the ERB in `config/database.yml` to fallback to a default
  username and pass if the env var isn't set.
- Require env vars in production. We don't want to make it easy to
  mistakenly use config we've published in our open source code.
- Provide the same defaults in our `docker-compose.yml`

Fixes #108