Add Redis to docker-compose.yml
The project depends on redis, but redis was not present in
`docker-compose.yml` Including it there is easy, and means that a single
`docker-compose up` is sufficient to have a working development and test
environment.

Also, I realized I introduced a bug in
https://github.com/opensourcecatholic/marriage-booklet/pull/110. I need
to quote my `database.yml` since they're strings.